Version of the CAN reader w/ Interrupts
The Battery communicates at 1000KBPS on CAN and thus if you're writing (Serial.println) to the serial port there is a loss of frames.  The MCP2515_CAN only has a 2 frame buffer, so if new frames arrive and the prior frame hasn't been processed (captured by the microcontroller) then the prior frame is overwritten and lost. The core problem is that data is coming in from the BMS faster than the microcontroller can output the frames to the serial port.

This test routine "CANTestToolInterrupt" uses two techniques to get around the problem, 1) interrupts and 2) circular queue.  The MCP2515 interrupt line is connected to pin 2 on the Mega, when the line goes LOW it triggers the interrupt data capture routine.  This routine just stuffs the captured frame into an array (it makes no attempt to output it to serial).  The circular queue has two pointers into it, the addPoint and the readPoint.  The interrupt writes the frames to the addPoint and the loop() reads frames for outputting to Serial from the readPoint.  The reader avoids reading if the addPoint and the readPoint are the same value.  The pointers cycle off the end of the array back to the start (thus circular).  Since the data is being added to the circular queue faster than it is being consumed, there are occasions where a frame will get overwritten by a new frame before the reader gets to it.  This is okay in the case of the BMS because the BMS just continuously blasts out the same frames repeatedly.  So if missed on one pass, the data can be caught on the next cycle.  This would be more problematic is messages were sent once and acted as commands, rather than informative (but that can also be handled at a higher level by sending a command (repeatedly) until a confirmation is received.